Reworked components structure for EventsPage
