+++ client/src/components/event-card.js
@@ -6,7 +6,7 @@ class EventCard extends React.Component {
 
   render() {
     return (
-      <Card className="eventCard m-2" border="dark" style={{opacity: '8'}}>
+      <Card className="eventCard m-2" border="dark">
         <Card.Img variant="top" src="/img/event-stub.svg" alt="Event image"/>
         <Card.Body>
           <Link to="#/"><Card.Title className="">Event Name</Card.Title></Link>
