--- client/src/pages/EventsPage/style.css
@@ -1,43 +1,0 @@
-.page-container{
-  background-image: url('/img/events-bkg.jpg');
-  background-size: cover;
-}
-
-.page-header-title{
-  padding-top: 2%;
-  color: #fac564;
-  text-align: center;
-  font-weight: bold;
-}
-
-.search-container{
-  padding: 0 15% 0 15%;
-}
-
-/* TODO: Style for buttons, rework later*/
-.m-button {
-  background: #fac564;
-  border-color: #fac564;
-  color: #000000;
-}
-
-.m-button:hover {
-  background: none;
-  border-color: #fac564;
-}
-
-.m-button:focus {
-  background: none !important;
-  border-color: #fac564 !important;
-  box-shadow: 0 0 0 0 !important;
-}
-
-.m-button:active{
-  background: #fac564 !important;
-  border-color: #fac564 !important;
-  box-shadow: 0 0 0 0 !important;
-}
-
-.react-datepicker__input-container input {
-  width: 75%;
-}
