+++ client/src/style/dropdown.css
@@ -1,0 +1,20 @@
+.m-dropdown-item:hover{
+  background:  #fac564;
+}
+
+.show>.btn-primary.dropdown-toggle{
+  background-color: #fac564;
+  border-color: #fac564;
+}
+
+.show>.btn-primary.dropdown-toggle:focus{
+  box-shadow: 0 0 0 0;
+  background-color: #fac564;
+  border-color: #fac564;
+}
+
+.show>.btn-primary{
+  box-shadow: 0 0 0 0;
+  background-color: #fac564;
+  border-color: #fac564;
+}
